requirements.txt
aniso8601==7.0.0
attrs==19.3.0
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
Django==2.2.10
django-dynamic-fixture==2.0.0
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.0.0
django-graphiql==0.4.4
django-graphql-jwt==0.1.5
django-model-utils==4.0.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.8
importlib-metadata==1.6.0
iso8601==0.1.12
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
promise==2.3
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.3.2
pytest-django==3.7.0
pytest-mock==1.13.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.9
wcwidth==0.2.2
zipp==3.1.0
